requirements.txt
ariadne-codegen==0.12.0
black==24.3.0
isort==5.13.2
pre-commit==3.6.0
pyjwt==2.8.0
ruff==0.3.5
